Fix invalid file links for rails best practices
Rails best practices returns absolute filenames, Template#complete_file_path is stripping off the leading slash of a file before expanding the path. This results in an invalid path when an absolute path is given; the working directory is appended to the already absolute path.
Removing the call to and definition of Template#remove_leading_slash fixes this issue.