Ruby 2.2 support
Ruby 2.2 is throwing a 'can't modify frozen NilClass' error on spec/support/suite.rb:11:in `artifact_test_dir'.
In artifact_test_dir an instance variable is being set on the current object, but at the end of the test suite this seems be be an empty (nil) context. In ruby 2.2 the nil object is frozen, so no instance variables can't be added to it anymore.
Removing the instance variable assignment fixes this, it's only caching one in two requests anyway.